Avoid circular import of `Beam2d` class
